# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BandFy",
        "description": "Sistema de comunicação e venda de serviços musicais",
        "version": "1.0.0"
    },
    "paths": {
        "/api/users/register": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Registrar um novo usuário",
                "description": "Cria um novo usuário na aplicação.",
                "operationId": "13565b949a19a5836c3fe9f30a37dbc6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "fullname",
                                    "main_phone",
                                    "optional_phone",
                                    "email",
                                    "password",
                                    "password_confirmed"
                                ],
                                "properties": {
                                    "fullname": {
                                        "type": "string",
                                        "example": "João Pé de Feijão"
                                    },
                                    "main_phone": {
                                        "type": "integer",
                                        "example": "61999999999"
                                    },
                                    "optional_phone": {
                                        "type": "integer",
                                        "example": "6177777777"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "joao@email.com.br"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "G@bri&L123"
                                    },
                                    "password_confirmed": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuário registrado com sucesso."
                    },
                    "422": {
                        "description": "Senhas não conferem."
                    },
                    "409": {
                        "description": "Usuário já cadastrado."
                    },
                    "500": {
                        "description": "Erro interno do servidor."
                    }
                }
            }
        },
        "/api/users/login": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Logar no sistema",
                "description": "Usuário pode fazer login e entrar no sistema",
                "operationId": "b5c0481a14cf25e2840c6c9dd1e4f3d0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "joao@email.com.br"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "******"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login realizado com sucesso."
                    },
                    "401": {
                        "description": "Credenciais inválidas."
                    },
                    "500": {
                        "description": "Erro interno do servidor."
                    }
                }
            }
        },
        "/api/users/logout": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Deslogar do sistema",
                "description": "Nesse endpoint o usuário irá deslogar de sua conta no sistema.",
                "operationId": "c3a125fe0fc085e4b3567e954639d4a0",
                "responses": {
                    "204": {
                        "description": "Logout realizado com sucesso. Nenhum conteúdo retornado."
                    },
                    "401": {
                        "description": "Credenciais inválidas"
                    },
                    "500": {
                        "description": "Erro interno do servidor."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/google/redirect": {
            "get": {
                "tags": [
                    "User - OAuth"
                ],
                "summary": "Redirect to Google OAuth",
                "description": "Redirects the user to Google OAuth for authentication.",
                "operationId": "redirectOAuth",
                "responses": {
                    "302": {
                        "description": "Redirect to Google OAuth",
                        "headers": {
                            "Location": {
                                "description": "Redirect location",
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Error message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/google/callback": {
            "get": {
                "tags": [
                    "User - OAuth"
                ],
                "summary": "Handle Google OAuth Callback",
                "description": "Handles the callback from Google after authentication and logs in the user.",
                "operationId": "oAuth",
                "responses": {
                    "302": {
                        "description": "Redirect to the dashboard after successful login",
                        "headers": {
                            "Location": {
                                "description": "Redirect location",
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Error message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Autenticação",
            "description": "Autenticação"
        },
        {
            "name": "User - OAuth",
            "description": "User - OAuth"
        }
    ]
}